requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
APScheduler==3.10.4
python-dotenv==1.0.1
Werkzeug==3.0.1
SQLAlchemy==2.0.28
numpy==1.26.4
scikit-learn==1.4.1.post1
python-Levenshtein==0.25.0
openai==1.12.0  # Optional for AI-powered matching
Authlib==1.3.0
requests==2.31.0 